requirements.txt
streamlit
langchain-openai
langchain-community
openai
python-dotenv
apify-client
requests
google-search-results